HU-03 Se ajusta la navegación
