+++ app/src/main/java/com/example/appmoviles/network/NetworkServiceAdapter.kt
@@ -9,6 +9,7 @@ import com.android.volley.VolleyError
 import com.android.volley.toolbox.StringRequest
 import com.android.volley.toolbox.Volley
 import com.example.appmoviles.models.Album
+import com.example.appmoviles.models.Performer
 import org.json.JSONArray
 
 class NetworkServiceAdapter constructor(context: Context) {
@@ -41,6 +42,22 @@ class NetworkServiceAdapter constructor(context: Context) {
                 onError(it)
             }
         ))
+    }
+
+    fun getPerformers(onComplete:(resp:List<Performer>)->Unit, onError: (error:VolleyError)->Unit){
+        requestQueue.add(getRequest("musicians",
+            Response.Listener<String> { response ->
+                val resp = JSONArray(response)
+                val list = mutableListOf<Performer>()
+                for (i in 0 until resp.length()) {
+                    val item = resp.getJSONObject(i)
+                    list.add(i, Performer(performerId = item.getInt("id"),name = if (item.has("name")) item.getString("name") else "", image = if (item.has("image")) item.getString("image") else "", description = if (item.has("description")) item.getString("description") else "", birthDate = if (item.has("birthDate")) item.getString("birthDate") else "", creationDate = if (item.has("creationDate")) item.getString("creationDate") else "", type = if (item.has("type")) item.getString("type") else "", bandId = if (item.has("bandId")) item.getInt("bandId") else 0))
+                }
+                onComplete(list.sortedBy { it.name })
+            },
+            Response.ErrorListener {
+                onError(it)
+            }))
     }
 
     private fun getRequest(path: String, responseListener: Response.Listener<String>, errorListener: Response.ErrorListener): StringRequest {
