+++ app/src/main/java/com/example/appmoviles/ui/adapters/PerformersAdapter.kt
@@ -1,8 +1,7 @@
-package com.example.appmoviles.adapters
+package com.example.appmoviles.ui.adapters
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import android.widget.ImageView
 import androidx.annotation.LayoutRes
 import androidx.databinding.DataBindingUtil
 import androidx.recyclerview.widget.RecyclerView
