+++ app/src/main/java/com/example/appmoviles/ui/performer/PerformerActivity.kt
@@ -1,0 +1,19 @@
+package com.example.appmoviles.ui.performer
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentTransaction
+import com.example.appmoviles.R
+
+class PerformerActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_performer_list)
+
+        supportActionBar!!.setDisplayHomeAsUpEnabled(true)
+
+    }
+
+}
