+++ app/src/main/res/layout/activity_performer_list.xml
@@ -4,32 +4,14 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="MainActivity">
-
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/my_toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="?attr/actionBarSize"
-        android:background="?attr/colorPrimary"
-        android:elevation="4dp"
-        android:textAlignment="center"
-        android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
-        app:layout_constraintTop_toTopOf="parent"
-        app:navigationIcon="@drawable/back_icon"
-        app:popupTheme="@style/ThemeOverlay.AppCompat.Light"
-        app:title="@string/artistasCapital"
-        app:titleTextColor="#FFFFFF"
-        tools:layout_editor_absoluteX="0dp" />
+    tools:context="ui.performer.PerformerActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/nav_host_fragment"
-        android:name="androidx.navigation.fragment.NavHostFragment"
+        android:name="com.example.appmoviles.ui.performer.PerformerFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_marginTop="64dp"
         app:defaultNavHost="true"
-        app:layout_constraintTop_toBottomOf="@+id/my_toolbar"
-        app:navGraph="@navigation/nav_graph"
         tools:layout_editor_absoluteX="50dp" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
