+++ app/src/main/res/layout/performer_fragment.xml
@@ -10,9 +10,9 @@
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.PerformerFragment"
-        android:layout_marginStart="16dp"
-        android:layout_marginEnd="16dp">
+        tools:context=".ui.performer.PerformerFragment"
+        android:layout_marginStart="5dp"
+        android:layout_marginEnd="5dp">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/fragments_rv"
