+++ app/src/main/res/layout/performer_item.xml
@@ -25,12 +25,14 @@
                 android:layout_height="80dp"
                 android:orientation="horizontal">
 
-                <ImageView
+                <de.hdodenhof.circleimageview.CircleImageView
                     android:id="@+id/performer_image"
                     android:layout_width="60dp"
                     android:layout_height="60dp"
                     android:gravity="center|center"
                     android:layout_margin="15dp"
+                    app:civ_border_width="2dp"
+                    app:civ_border_color="#FF000000"
                     />
 
                 <TextView
