+++ local.properties
@@ -7,4 +7,4 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-sdk.dir=/Users/juansebastianb/Library/Android/sdk
+sdk.dir=C\:\\Users\\NICOLAS\\AppData\\Local\\Android\\Sdk
